+++ packages/common/src/compose-tunnel-agent.ts
@@ -1,4 +1,4 @@
-import { camelCase, set, snakeCase } from 'lodash'
+import { extractSectionsFromLabels, parseBooleanLabelValue } from './compose-utils'
 
 export const COMPOSE_TUNNEL_AGENT_SERVICE_LABELS = {
   PROFILE_THUMBPRINT: 'preevy.profile_thumbprint',
@@ -22,6 +22,7 @@ export type ScriptInjection = {
 type Stringified<T> = {
   [k in keyof T]: string
 }
+
 const parseScriptInjection = ({ pathRegex, defer, async, src }: Stringified<ScriptInjection>):
  ScriptInjection | Error => {
   try {
@@ -30,25 +31,13 @@ const parseScriptInjection = ({ pathRegex, defer, async, src }: Stringified<Scri
     }
     return {
       ...pathRegex && { pathRegex: new RegExp(pathRegex) },
-      ...defer && { defer: defer === 'true' },
-      ...async && { async: async === 'true' },
+      ...defer && { defer: parseBooleanLabelValue(defer) },
+      ...async && { async: parseBooleanLabelValue(async) },
       src,
     }
   } catch (e) {
     return e as Error
   }
-}
-
-export const extractSectionsFromLabels = <T>(prefix: string, labels: Record<string, string>) => {
-  const re = new RegExp(`^${prefix.replace(/\./g, '\\.')}\\.(?<id>.+?)\\.(?<key>[^.]+)$`)
-  const sections:{[id:string]: T } = {}
-  for (const [label, value] of Object.entries(labels)) {
-    const match = label.match(re)?.groups
-    if (match) {
-      set(sections, [match.id, camelCase(match.key)], value)
-    }
-  }
-  return sections
 }
 
 export const scriptInjectionFromLabels = (labels : Record<string, string>): ScriptInjection[] => {
@@ -60,13 +49,3 @@ export const scriptInjectionFromLabels = (labels : Record<string, string>): Scri
     .map(parseScriptInjection)
     .filter((x): x is ScriptInjection => !(x instanceof Error))
 }
-
-const formatValueLabel = (x:unknown) => {
-  if (x instanceof RegExp) {
-    return x.source
-  }
-  return `${x}`
-}
-
-export const sectionToLabels = (prefix: string, section: Record<string, unknown>) =>
-  Object.fromEntries(Object.entries(section).map(([key, value]) => ([`${prefix}.${snakeCase(key)}`, formatValueLabel(value)])))
